+++ spec/generators/rails_template18f/cloud_gov_config/cloud_gov_config_generator_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe RailsTemplate18f::Generators::CloudGovConfigGenerator, type: :gen
   before { run_generator }
 
   it "adds the ClimateControl gem" do
-    expect(file("Gemfile")).to contain('gem "climate_control", "~> 1.0"')
+    expect(file("Gemfile")).to contain('gem "climate_control", "~> 1.2"')
   end
 
   it "copies the CloudGovConfig model and test" do