+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Set up app space via terraform, with proper restricted egress security group
 - Create rails_template18f:public_egress generator for integrating with cg-egress-proxy
 - [Use exec when starting rails server](https://docs.cloudfoundry.org/devguide/deploy-apps/manifest-attributes.html#start-commands:~:text=To%20resolve%20this,process.%20For%20example%3A)
+- Upgrade the i18n-js integration to 4.x
 
 ## [1.2.0] - 2024-09-20